+++ src/main/java/verbly/spring/domain/post/controller/post/PostControllerDocs.java
@@ -1,4 +1,4 @@
-package verbly.spring.domain.post.controller;
+package verbly.spring.domain.post.controller.post;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
